refactor: 응답 dto로 매핑하는 로직을 mapper로 이동
