+++ src/main/java/com/dmz/airdnd/accommodation/mapper/AccommodationMapper.java
@@ -9,6 +9,7 @@ import com.dmz.airdnd.accommodation.domain.Label;
 import com.dmz.airdnd.accommodation.dto.response.AccommodationResponse;
 import com.dmz.airdnd.accommodation.dto.response.AddressResponse;
 import com.dmz.airdnd.accommodation.dto.response.LabelResponse;
+import com.dmz.airdnd.accommodation.dto.response.AccommodationCreateResponse;
 
 public class AccommodationMapper {
 	public static Accommodation toEntity(AccommodationCreateRequest request, Address address) {
@@ -54,5 +55,30 @@ public class AccommodationMapper {
 
 	public static List<LabelResponse> toResponses(List<Label> labels) {
 		return labels.stream().map(label -> new LabelResponse(label.getId(), label.getName())).toList();
+	}
+
+	public static AccommodationCreateResponse fromEntity(Accommodation accommodation) {
+		return AccommodationCreateResponse.builder()
+			.id(accommodation.getId())
+			.address(formatFullAddress(accommodation.getAddress()))
+			.name(accommodation.getName())
+			.description(accommodation.getDescription())
+			.pricePerDay(accommodation.getPricePerDay())
+			.currency(accommodation.getCurrency())
+			.maxGuests(accommodation.getMaxGuests())
+			.bedCount(accommodation.getBedCount())
+			.bedroomCount(accommodation.getBedroomCount())
+			.bathroomCount(accommodation.getBathroomCount())
+			.createdAt(accommodation.getCreatedAt())
+			.build();
+	}
+
+	private static String formatFullAddress(Address address) {
+		String base = address.getBaseAddress();
+		String detail = address.getDetailedAddress();
+		if (detail != null && !detail.isEmpty()) {
+			return base + " " + detail;
+		}
+		return base;
 	}
 }
